Pushing image to datastore and deleting after inference (#4)
* Pushing image to datastore

And immediately deleting 'local' copy (from the snapshot)

* Deleting image file after bootstar script runs

Not tested yet.

* two spurious newlines

* mypy spotted missing colon

* better defaults

* Fixing two mypy errors

* Test prints

Using AzureML Studio to work out is shapshot is affected by the deletion

* Saves image to datastore

* Including GUID in datastore path

* WiP downloading images from datastore in bootstrap script

* Overwriting image data zip

* tidy up

* Returning GUID and tidy-up

* Added unit test of image data zip overwrite

* Fixing API call to submit with ignored return

* Swapping temp image folder name to config

* Fixing documentation

* Plumbing for datastore name

* Swapping from default to named datastore

* Fixing main.yml build env

* Decreasing indentation by breaking loop

https://github.com/microsoft/InnerEye-Inference/pull/4#discussion_r617584432

* pylint not used and commented out line deleted

* Fixing flake8 warnings

* Swapping imports back to single line

After Anton's query on the PR I checked the .flake8 file and our maximum line length is 160 and
so I do not need to break up these lines.

* Hard coding not-secret settings in workflow

As per Anton's request:
https://github.com/microsoft/InnerEye-Inference/pull/4#discussion_r618144366

* /= consistency

* removing 2 debug print statements

* removed brackets

* swapping to run.wait_for_completion

* swapping to writetext

* removing unnecessary initialization

* Rationalising 'run' method: paths, arguments, and comments

* removing debug print lines

* Reverting line lengths

I had assumed a max line length of 100, but it is 160 in the .flake8 configuration file so
I have reverted the changes I had made to not-too-long lines, and fixed a few others for
legibility

* Swapping to required=True for params

* Typo and unnecessary Path() spotted by Jonathan

* Changing parameter to underscores from dashes

To bring them into line with the rest of the InnerEye projects. They needed
to be different when the unknown args were passed straight on to score.py
but that is not how the arguments flow through anymore.

* Line length fix in comment to kick off license/cla

The 'license/cla' check in the build pipeline has stalled and there
is not 'retry' buttong so I am hoping to kick it into action again
with an almost vacuaous commit!